requirements.txt
# TOX requirements
-r requirements-tox.txt

# Minimum Django and REST framework version
Django>=1.11
djangorestframework>=3.6.0